chore(e2e): reorg and clean up e2e/global/builder
